+++ 2021/2sms/SO/p2/complex/complex.h
@@ -1,17 +1,17 @@
 /* definition of new type complex */
 typedef struct {
-	double x;
-	double y;
+	int x;
+	int y;
 } complex;
 
 /* definition of the complex API */
-complex* complex_new(double, double);
+complex* complex_new(int, int);
 complex* complex_add(complex *, complex *);
 complex* complex_sub(complex *, complex *);
 complex* complex_mul(complex *, complex *);
 complex* complex_div(complex *, complex *);
 complex* complex_conj(complex *);
-double complex_mod(complex *);
-double complex_arg(complex *);
-double complex_re(complex *);
-double complex_im(complex *);
+int complex_mod(complex *);
+int complex_arg(complex *);
+int complex_re(complex *);
+int complex_im(complex *);
